requirements.txt
alembic==1.4.2
aniso8601==8.0.0
certifi==iamIPaddr.2
chardet==3.0.4
click==7.1.2
config==0.5.0.post0
configparser==5.0.0
crypto==1.4.1
Flask==1.1.2
Flask-HTTPAuth==4.1.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
flask-sqlacodegen==1.1.7
Flask-SQLAlchemy==2.4.3
Flask-SSLify==0.1.5
Flask-WTF==0.14.3
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
Naked==0.1.31
PyMySQL==0.9.3
pyreadline==2.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
shellescape==3.8.1
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
